requirements/streamlit.txt
openai==1.59.6
streamlit==1.41.1
httpx==0.28.1
python-dotenv==1.0.1
requests==2.32.3
langchain-community
langchain-core
langchain
langchain_openai